# static/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API to interact with [Hackatime](https://w.dunkirk.sh)\n\n## Authentication\nSet header `Authorization` to your API Key encoded as Base64 and prefixed with `Basic`\n**Example:** `Basic ODY2NDhkNzQtMTljNS00NTJiLWJhMDEtZmIzZWM3MGQ0YzJmCg==`\n\nIf you are using the admin token then set header `Authorization` to the admin token prefixed with `Bearer` and not encoded with anything\n**Example:** `Bearer blahaji_rulz_da_world` One point to be aware of is that when using this method you always need to pass the `user` query parameter with the id of the user you wish to query for.",
        "title": "Hackatime API",
        "contact": {
            "name": "Kieran Klukas",
            "url": "https://github.com/kcoderhtml",
            "email": "me@dunkirk.sh"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/kcoderhtml/hackatime/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Retrieve a summary",
                "operationId": "get-summary",
                "parameters": [
                    {
                        "enum": [
                            "today",
                            "yesterday",
                            "week",
                            "month",
                            "year",
                            "7_days",
                            "last_7_days",
                            "30_days",
                            "last_30_days",
                            "6_months",
                            "last_6_months",
                            "12_months",
                            "last_12_months",
                            "last_year",
                            "any",
                            "all_time",
                            "low_skies"
                        ],
                        "type": "string",
                        "description": "Interval identifier",
                        "name": "interval",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (e.g. '2021-02-07')",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (e.g. '2021-02-08')",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to recompute the summary from raw heartbeat or use cache",
                        "name": "recompute",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Project to filter by",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Language to filter by",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Editor to filter by",
                        "name": "editor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "OS to filter by",
                        "name": "operating_system",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Machine to filter by",
                        "name": "machine",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Project label to filter by",
                        "name": "label",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The user to filter by if using Bearer authentication and the admin token",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Summary"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Summary": {
            "type": "object",
            "properties": {
                "branches": {
                    "description": "branches are not persisted, but calculated at runtime in case a project Filter is applied",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "editors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "entities": {
                    "description": "entities are not persisted, but calculated at runtime in case a project Filter is applied",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02 15:04:05.000"
                },
                "labels": {
                    "description": "labels are not persisted, but calculated at runtime, i.e. when summary is retrieved",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "operating_systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "projects": {
                    "description": "Previously, all the following properties created a cascade foreign key constraint on the summary_items table\nback to this summary table resulting in 5 identical foreign key constraints on the summary_items table.\nThis is not a problem for PostgreSQL, MySQL and SQLite, but for MSSQL, which complains about circular cascades on\nupdate/delete between these two tables. All of these created foreign key constraints are identical, so only one constraint is enough.\nMySQL will create a foreign key constraint for every property referencing other structs, even no constraint is specified in tags.\nSo explicitly set gorm:\"-\" in all other properties to avoid creating duplicate foreign key constraints",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "to": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02 15:04:05.000"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.SummaryItem": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}